pyproject.toml
[project]
name = "azure-ai-foundry-agents-deep-research"
version = "1.0.0"
description = "Azure AI Deep Research application with modular architecture following SOLID principles"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "azure-ai-agents==1.2.0b1",
    "azure-ai-projects>=1.0.0",
    "azure-core-tracing-opentelemetry>=1.0.0b12",
    "azure-identity>=1.24.0",
    "azure-monitor-opentelemetry>=1.6.13",
    "opentelemetry-exporter-otlp>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.40.0",
    "watchdog>=6.0.0",
    "dataclasses-json>=0.6.3",
    "typing-extensions>=4.5.0",
    "pydantic>=2.0.0",
    "bleach>=6.0.0",
]

[project.scripts]
azure-ai-research = "azure_ai_research.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[tool.setuptools.packages.find]
include = ["azure_ai_research*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
